==> Notebooks/02_Preparing_Datasets.ipynb <==
# Libraries
import scanpy as sc
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import deconomix as dcx
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)

# DISCO Dataset: `disco_breast_v01.h5ad` from https://zenodo.org/records/7396984.
data = sc.read_h5ad('../Data/DISCO/disco_breast_v01.h5ad', backed='r')

# Import consensus gene set, determined in B01_Consensus_Gene_Selection.ipynb
genes_and_conv = pd.read_csv('../Data/Preprocessed/final_genes_and_conv.csv')
consensus_genes = genes_and_conv.loc[:,"gene"].values.flatten()

ct_aggregation = pd.read_csv('../Data/DISCO/cell_type_aggregation.csv')
ct_major = dict(zip(ct_aggregation['Cell type (DISCO)'], ct_aggregation['Cell type (major)']))
ct_minor = dict(zip(ct_aggregation['Cell type (DISCO)'], ct_aggregation['Cell type (minor)']))

# Filtering for solid breast cancer samples
data_cancer = data[(data.obs['disease'] == 'breast cancer') & (data.obs['sample_type'] == 'solid tumor')]
data_healthy = data[(data.obs['disease'] == 'NA') & (data.obs['sample_type'] == 'normal')]
# Attach original ct labels
ct_labels_cancer = data_cancer.obs['ct']
ct_labels_healthy = data_healthy.obs['ct']
# Choose consensus genes
gene_mask = data.var.index.isin(consensus_genes)

disco_cancer = pd.DataFrame(data_cancer.X[:,gene_mask].toarray().T,
                        index=data_cancer.var['gene'][gene_mask].values,
                        columns=ct_labels_cancer)

#Normalize to 10k counts per sample
disco_cancer_norm = disco_cancer * (10000 / disco_cancer.sum())

# Reorder index to match order of consensus genes
disco_cancer = disco_cancer.loc[consensus_genes, :]
disco_cancer_norm = disco_cancer_norm.loc[consensus_genes, :]

# Save results to pickle
disco_cancer.to_pickle('../Data/Preprocessed/disco_cancer.pkl')
disco_cancer_norm.to_pickle('../Data/Preprocessed/disco_cancer_norm.pkl')

disco_healthy = pd.DataFrame(data_healthy.X[:,gene_mask].toarray().T,
                        index=data_healthy.var['gene'][gene_mask].values,
                        columns=ct_labels_healthy)

# drop cancer epithelial cells
disco_healthy = disco_healthy.loc[:, disco_healthy.columns.map(ct_major) != 'Cancer Epithelial']

# Normalize to 10k counts per sample
disco_healthy_norm = disco_healthy * (10000 / disco_healthy.sum())

# Reorder index to match order of consensus genes
disco_healthy = disco_healthy.loc[consensus_genes, :]
disco_healthy_norm = disco_healthy_norm.loc[consensus_genes, :]

# Save results to pickle
disco_healthy.to_pickle('../Data/Preprocessed/disco_healthy.pkl')
disco_healthy_norm.to_pickle('../Data/Preprocessed/disco_healthy_norm.pkl')


# TCGA data downloaded by running tcga_downloader.py once, extract the downloaded archives, then run it again...
data = pd.read_csv("../Data/TCGA/tcga_brca_rnaseq_counts.csv", index_col=0)
pam50_row = data.loc["PAM50"].tolist()  # Extract PAM50 row as a list
data = data.drop("PAM50")  # Remove PAM50 row from main data
data = data.apply(pd.to_numeric, errors="coerce").astype("Int64")
pd.Series(pam50_row).to_csv('../Data/Preprocessed/tcga_subtype.csv', index=False)

# select consensus genes and scale bulks with conversion factors

data

# select consensus genes and apply conversion factors
conv_factors = genes_and_conv.loc[:, "conversion factor"].values.flatten()
tcga_conv = data.loc[consensus_genes, :].multiply(conv_factors, axis=0)
tcga_conv

tcga_conv_norm = tcga_conv * (10000/ tcga_conv.sum())
tcga_conv_norm

tcga_conv.to_pickle('../Data/Preprocessed/tcga_conv.pkl')
tcga_conv_norm.to_pickle('../Data/Preprocessed/tcga_conv_norm.pkl')
