==> Notebooks/03_DISCO_Cancer-DISCO_Cancer.ipynb <==
# Libraries
import scanpy as sc
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import deconomix as dcx
from scipy.stats import spearmanr, pearsonr
from eval_functions import create_composition_overview, create_performance_overview

# Ct label aggregation
ct_aggregation = pd.read_csv('../Data/DISCO/cell_type_aggregation.csv')
ct_major = dict(zip(ct_aggregation['Cell type (DISCO)'], ct_aggregation['Cell type (major)']))
ct_minor = dict(zip(ct_aggregation['Cell type (DISCO)'], ct_aggregation['Cell type (minor)']))
ct_aggregation

# Data = Disco Cancer
data = pd.read_pickle('../Data/Preprocessed/disco_cancer_norm.pkl')
# choose minor cts
data.columns = data.columns.map(ct_minor)
data = data.rename(columns={'Cancer Epithelial': 'hidden'})

# Train/Test split
n_obs = data.shape[1]
np.random.seed(42)
train_test_split = np.random.choice([0, 1], size=n_obs, p=[0.5, 0.5])

data_train = data.loc[:,train_test_split==0].drop('hidden', axis=1)
data_test = data.loc[:,train_test_split==1]

table_S1 = create_composition_overview(data_train, data_test)
table_S1.to_latex(buf="../Outputs/Table_S1.tex",
                  index=False,
                  na_rep='-',
                  float_format='%.3f',
                  column_format='l|c|c|c|c')
table_S1

table_S2 = create_performance_overview(data_train, data_test)
table_S2.to_latex(buf="../Outputs/Table_S2.tex",
                  index=False,
                  na_rep='-',
                  float_format='%.3f',
                  column_format='l|c|c|c|c|c')
table_S2
