==> Notebooks/04_DISCO_Healthy-DISCO_Cancer.ipynb <==
# Libraries
import scanpy as sc
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import deconomix as dcx
from scipy.stats import spearmanr, pearsonr
from eval_functions import create_composition_overview, create_performance_overview

# Ct label aggregation
ct_aggregation = pd.read_csv('../Data/DISCO/cell_type_aggregation.csv')
ct_major = dict(zip(ct_aggregation['Cell type (DISCO)'], ct_aggregation['Cell type (major)']))
ct_minor = dict(zip(ct_aggregation['Cell type (DISCO)'], ct_aggregation['Cell type (minor)']))
#ct_aggregation

# Data Import
data_cancer = pd.read_pickle('../Data/Preprocessed/disco_cancer_norm.pkl')
data_cancer.columns = data_cancer.columns.map(ct_major)
data_cancer = data_cancer.rename(columns={'Cancer Epithelial': 'hidden'})
data_healthy = pd.read_pickle('../Data/Preprocessed/disco_healthy_norm.pkl')
data_healthy.columns = data_healthy.columns.map(ct_major)

table_S5 = create_composition_overview(data_healthy, data_cancer)
table_S5.to_latex("../Outputs/Table_S5.tex", index=False, na_rep='-', float_format='%.3f', column_format='l|c|c|c|c')
table_S5

table_S6 = create_performance_overview(data_healthy, data_cancer)
table_S6.to_latex("../Outputs/Table_S6.tex", index=False, na_rep='-', float_format='%.3f', column_format='l|c|c|c|c|c')
table_S6
