==> Notebooks/05a_Cache_Artificial_Bulks.ipynb <==
# Libraries
import scanpy as sc
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import deconomix as dcx
from scipy.stats import spearmanr, pearsonr
from eval_functions import create_composition_overview, create_performance_overview

# Ct label aggregation DISCO
ct_aggregation = pd.read_csv('../Data/DISCO/cell_type_aggregation.csv')
ct_major = dict(zip(ct_aggregation['Cell type (DISCO)'], ct_aggregation['Cell type (major)']))
ct_minor = dict(zip(ct_aggregation['Cell type (DISCO)'], ct_aggregation['Cell type (minor)']))

# Data Import
print("Importing Data")
disco_healthy = pd.read_pickle('../Data/Preprocessed/disco_healthy_norm.pkl')
#tcga_bulks   = pd.read_pickle('Data/Preprocessed/tcga_conv_norm.pkl')
#subtype = pd.read_csv('Data/Preprocessed/tcga_subtype.csv').values.flatten()

# Select major labels
disco_healthy.columns = disco_healthy.columns.map(ct_major)

# Generate training data from disco sc reference
print("Simulating Data")
X_ref, Y_train, C_train = dcx.utils.simulate_data(disco_healthy, n_mixtures=20000, n_cells_in_mix=100)
X_ref.to_pickle('../Data/Preprocessed/X_ref.pkl')
Y_train.to_pickle('../Data/Preprocessed/Y_train.pkl')
C_train.to_pickle('../Data/Preprocessed/C_train.pkl')
