==> Notebooks/07_Downstream_Kruskal-Wallis.ipynb <==
# Libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import matplotlib as mpl

# Data Import
PREFIX = '05c_l1e-6_'
C_LumA = pd.read_csv(f"../Outputs/{PREFIX}C_est_LumA.csv", index_col=0)
C_LumB = pd.read_csv(f"../Outputs/{PREFIX}C_est_LumB.csv", index_col=0)
C_Her2 = pd.read_csv(f"../Outputs/{PREFIX}C_est_Her2.csv", index_col=0)
C_Basal = pd.read_csv(f"../Outputs/{PREFIX}C_est_Basal.csv", index_col=0)
c_LumA = pd.read_csv(f"../Outputs/{PREFIX}c_est_LumA.csv", index_col=0)
c_LumB = pd.read_csv(f"../Outputs/{PREFIX}c_est_LumB.csv", index_col=0)
c_Her2 = pd.read_csv(f"../Outputs/{PREFIX}c_est_Her2.csv", index_col=0)
c_Basal = pd.read_csv(f"../Outputs/{PREFIX}c_est_Basal.csv", index_col=0)




C_full_LumA = pd.concat((C_LumA, c_LumA), axis=0)
C_full_LumB = pd.concat((C_LumB, c_LumB), axis=0)
C_full_Her2 = pd.concat((C_Her2, c_Her2), axis=0)
C_full_Basal = pd.concat((C_Basal, c_Basal), axis=0)

c_LumA

cell_types = list(C_full_LumA.index)
groups = {"LumA": C_full_LumA, "LumB": C_full_LumB, "Her2": C_full_Her2, "Basal": C_full_Basal}

# Mean and std across samples (columns) for each cell type (row)
means = pd.DataFrame({name: df.mean(axis=1, skipna=True) for name, df in groups.items()})
stds  = pd.DataFrame({name: df.std(axis=1, ddof=1, skipna=True) for name, df in groups.items()})

# Format as "mean ± std"
overview = means.round(3).map(lambda x: f"{x:.3f}").astype(str) + " ± " + stds.round(3).map(lambda x: f"{x:.3f}").astype(str)
overview.index.name = "cell_type"
overview


from statsmodels.stats.multitest import multipletests
from scipy.stats import kruskal

# List of cell types (assuming row index is the same across all DataFrames)

results = []
for ct in cell_types:
    arrays = []
    for group in groups:
        arrays.append(groups[group].loc[ct].values)
    H, p = kruskal(*arrays)
    results.append([ct, H, p])

results = pd.DataFrame(results, columns=["Cell Type", "H-statistic", "p-value"])




results.set_index("Cell Type", inplace=True)

# multiple testing correction
results["p-value-corrected"] = multipletests(results["p-value"], method="fdr_bh")[1]#.map(lambda x: f"{x:.3g}")
#results = results.sort_values(by="p-value-corrected")

results



final_table = pd.concat((overview, results), axis=1)
final_table

final_table.loc[:,"p-value"] = results["p-value"].astype(np.float64).map(lambda x: f"{x:.3g}")

final_table.loc[:,"p-value-corrected"] = results["p-value-corrected"].map(lambda x: f"{x:.3g}")

final_table.loc[:,"H-statistic"] = results["H-statistic"].astype(np.float64).map(lambda x: f"{x:.3f}")

final_table

final_table = final_table.reset_index(names="Cell type")

final_table.to_latex("../Outputs/Table_Kruskal-Wallis.tex", index=False, na_rep='-', column_format='l|c|c|c|c|c|c|c')
